Created the node structure
For this commit I created the "node" structure which contains an integer
called data.  Additionally  it contains a pointer to the next node.  In
the "main" I put the syntax of interacting with the node such as
creating a new node, updating the data, and creating and updating the
pointer.