Added more methods to C++ Singly Linked List
I noticed that my C++ linked list did not have all the methods I wanted.
I added some more methods and updated some names of the methods.
Specifically I created the displayFront() and displayend() methods which
display the first and last nodes in the linked list.  Additionally I
added removeFront() and removeEnd() methods which remove the first and
last nodes in the linked list.  Although there are theoretically more
methods I could add to this linked list I'm pretty content with it right
now.  My next goal is to create a doubly linked list with similar or the
same methods that this current linked list has.