Started Hash Table with C++
Started a Hash Table, started by making a simple linked list based on
the other singly linked list I've already posted to Github.
Unfortunately the "get" function is buggy and anytime I try to "get" a
value that isn't there the program crashes.  I've tried using "throw 1"
and other means of error handling and it doesn't work.  However since
the linked list portion of this code isn't my focus I'm going to simply
not use "get" when the linked lists are empty.  The other nodes work
fine.  I think the issue is that the RETURN statement needs to return a
variable of type T (template type) and when I throw an error it's not
returning anything.