Created a Binary Search Tree
I don't know if it's perfect.  But the few tests I've run it through
seem the work.  The biggest challenge by a HUGE gap was the delete
function.  Everything else took maybe a few minutes a piece.  The delete
function took me over an hour PLUS a lot of research.  I forgot how
complicated it was to rotate all those pointers.